split fixtures into own module
